split up/restructure API functions in UI
