+++ src/ui/server/api/index.ts
@@ -1,0 +1,9 @@
+import external from './external';
+import keystone from './keystone';
+
+const api = {
+  external,
+  keystone,
+};
+
+export default api;
