+++ src/ui/server/api/keystone/index.ts
@@ -1,0 +1,11 @@
+import application from './application';
+import country from './country';
+import page from './page';
+
+const keystone = {
+  application,
+  country,
+  page,
+};
+
+export default keystone;
