+++ src/ui/server/controllers/insurance/eligibility/buyer-country/index.test.ts
@@ -49,10 +49,10 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
 
     beforeEach(() => {
       delete req.session.submittedData.quoteEligibility[FIELD_IDS.BUYER_COUNTRY];
-      api.getCountries = getCountriesSpy;
+      api.external.getCountries = getCountriesSpy;
     });
 
-    it('should call api.getCountries', async () => {
+    it('should call api.external.getCountries', async () => {
       await get(req, res);
 
       expect(getCountriesSpy).toHaveBeenCalledTimes(1);
@@ -96,7 +96,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
       beforeEach(() => {
         // @ts-ignore
         getCountriesSpy = jest.fn(() => Promise.resolve());
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
@@ -109,7 +109,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
       beforeEach(() => {
         // @ts-ignore
         getCountriesSpy = jest.fn(() => Promise.resolve({}));
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
@@ -121,7 +121,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
     describe('when the CIS (country information system) API does not return a populated array', () => {
       beforeEach(() => {
         getCountriesSpy = jest.fn(() => Promise.resolve([]));
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
@@ -135,7 +135,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
     let getCountriesSpy = jest.fn(() => Promise.resolve(mockCountriesResponse));
 
     beforeEach(() => {
-      api.getCountries = getCountriesSpy;
+      api.external.getCountries = getCountriesSpy;
     });
 
     describe('when there are validation errors', () => {
@@ -244,7 +244,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
       beforeEach(() => {
         // @ts-ignore
         getCountriesSpy = jest.fn(() => Promise.resolve());
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
@@ -257,7 +257,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
       beforeEach(() => {
         // @ts-ignore
         getCountriesSpy = jest.fn(() => Promise.resolve({}));
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
@@ -269,7 +269,7 @@ describe('controllers/insurance/eligibility/buyer-country', () => {
     describe('when the CIS (country information system) API does not return a populated array', () => {
       beforeEach(() => {
         getCountriesSpy = jest.fn(() => Promise.resolve([]));
-        api.getCountries = getCountriesSpy;
+        api.external.getCountries = getCountriesSpy;
       });
 
       it(`should redirect to ${ROUTES.PROBLEM_WITH_SERVICE}`, async () => {
