+++ src/ui/server/controllers/quote/buyer-country/index.ts
@@ -48,7 +48,7 @@ export const getBackLink = (referer?: string): string => {
 
 export const get = async (req: Request, res: Response) => {
   const { submittedData } = req.session;
-  const countries = await api.getCountries();
+  const countries = await api.external.getCountries();
 
   if (!countries || !countries.length) {
     return res.redirect(ROUTES.PROBLEM_WITH_SERVICE);
@@ -80,7 +80,7 @@ export const get = async (req: Request, res: Response) => {
 export const post = async (req: Request, res: Response) => {
   const validationErrors = generateValidationErrors(req.body);
 
-  const countries = await api.getCountries();
+  const countries = await api.external.getCountries();
 
   if (!countries || !countries.length) {
     return res.redirect(ROUTES.PROBLEM_WITH_SERVICE);
